backend/requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.4.4
flask-cors==3.0.10
stripe==5.1.1
python-dotenv==1.0.0